calc: deal with late arrival of sheetGeometry message on document-load
On document load, it is possible (even if improbable) for the first
sheetGeometry message to arrive after other messages (like
cell-cursor/edit-cursor/selection etc) which need the former to be used
correctly. So, guard against crashing on the missing sheetGeometry
object in the docLayer (and let the print-twips coordinates in the
messages to be used for display which is incorrect). The incorrect
positions are corrected anyway because we replay those messages, when
the sheetGeometry message finally arrives.

Change-Id: I21840ca08ca62213cba3a4d5aa6e9462d11a5e0e