We don't check for ENABLE_SETCAP in any source file so no need to AC_DEFINE it
AC_DEFINE causes it to be in config.h, but there is no code that would
use its definition from there.

It is enough to have AM_CONDITIONAL for it (to enable having 'if
ENABLE_SETCAP' in Makefile.am files) and AC_SUBST it (to enable having
'@ENABLE_SETCAP@' in Makefile.am and *.in files).

Change-Id: Ia00b624114c8139d81bb173c92800ae0a62fec35