Update patch set 1
Patch Set 1:

I have a vague suspicion that it used to do this back in the day and we switched away from that; can you test on mobile, tablet etc. somehow ? how does it come with real screen-size vs. some presumed 'page' size that can be larger ?
Anyhow - lovely if it fixes things ! =)

Patch-set: 1
CC: Gerrit User 1000865 <1000865@6622c302-437a-49ad-a91d-55ba7dcf1615>
CC: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>