Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (2 comments)
> 
> Hi Matheus, first - thanks for contributing to Online ! =) I read through the changes. In general doing manual linting of this sort - is problematic. I'd love to help you move onto some real functional improvement instead - what are you interested in working on ? Thanks.

Thank you for your response, Michael! I didn't know that these sort of changes were problematic, the formatting stuck out to me as possibly amiss because it didn't follow the same conventions as the other files in the js folder. I also didn't know that the CustomScrollbar file was an external embed, sorry for that. I have experience with JavaScript and PHP (which, to my surprise, isn't used by Online at all!), plus the front-end side (HTML and CSS), so I would only be able to help with the loleaflet and cypress_test folders (I don't have experience with Cypress, however).

Patch-set: 1