Don't update modified status after saving to storage fails
Otherwise client gets a notification that document is unmodified.
This should not happen, as the document in the storage has not been updated
and so it should be considered as modified until saving to storage succeeds.

(cherry picked from commit dd87f6bf89e5e92918fb15265c5861d2e817286e)

Change-Id: I6918f97d96a546ce086f622854f4cbeed48d54ae