android: Make the "create new file" FAB always visible.
There were reports that it didn't appear after a new installation.
I failed to reproduce that, but I can imagine it might have happened,
seeing the code.  But it is not important to hide it any more, because
we are using the Content providers now.

When at this, make the FAB rotate the other direction, to look more
consistent with the appearing / disappearing buttons.

Change-Id: I92e01c7190405c2d4869edd50a9da011558306af