loleaflet: use the empty tile-container as a proxy...
... for the canvas to get the mouse-move+drag events and the various
touch events. This is needed because the canvas element (created by
L.CanvasTileLayer) is a sibling of the map-pane div and not a decendant.
Any change in canvas-size is also applied on the tile-container. Also if
there is a move event, counter adjust the tile-container position so
that it is precisely over the canvas at all times.

Change-Id: I5d5bc070c0063c5f24c4b1cb1448db0ce19841bd