split-panes: update svg container position and viewbox on 'drag'...
so that they do not get out of sync with the canvas on mobile/tablet
swipe/pan.

Note: This does not update the contents(path nodes) of the svg's.  Note:
In the long term we can get rid of the map-pane scrolling and avoid all
this but right now only Calc is using canvas based tile layer.

Change-Id: Iaf6894b0fc39a2a32ac19d0e9acd6abb14336785