Update patch set 6
Patch Set 6: Code-Review-1

What is this hack with the cypressTest flag? You changed the behavior of the back button and instead of updating the related tests, you hacked the code to keep the old behavior for cypress tests. Better to disable the failing tests instead, so we can see which tests need to be updated, if you do not intend to update those tests.

Patch-set: 6
Reviewer: Gerrit User 1000169 <1000169@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1