leaflet: hide the splitter on the toolbar if the first
When the toobar buttons to the left of the first splitter
are hidden, we start the toolbar with a splitter, which
is aesthetically unpleasant.

Here we special case for the desktop only, because
tablets and mobiles have a lightly different toolbar.
Although we should have a similar logic for them too.

Change-Id: I9869433f3ca3391d7ee899a77bb1466fe456f9f2