cypress: introduce typeIntoDocument helper method.
It was not consistent what we used to type into the
document (e.g. 'body', 'textarea.clipboard' or
'#document-container'). Also we need to use force
parameter because 'textarea.clipboard' has no actual area
which makes cypress fail with: '`cy.type()` failed because
the center of this element is hidden from view'.

Change-Id: I04e0a82ed8450d583e407715fde1855197cae109