Stopgap fix to make the on-screen keyboard behave a bit better on iPhone
Our focus and blur back-and-forth dance does not seem to work on iOS
like it does on Android. So for now, never call the textarea element's
blur() function explicitly on iOS.

Also don't play with the readonly attribute on iPhone.

Probably whether the calls to the textarea's focus() and blur()
functions actually do anything or not might depend on whether the call
stack originates in a user input event handler or not, for security
reasons.

Change-Id: I442f0c26ae5b36913235a7ef96820bff83535b76