cypress: improve removeShapeSelection() method.
Instead of using a click() and a dblclick() method, do
click() repeatedly, until selection is actually removed.
For example, it takes ony click when the shape is selected,
but no text is selected inside it. It takes two clicks when
a text also selected and sometimes when selection stucks it
takes even more click.

Change-Id: Ic017d0f5b623de29343cf171a8f4d64988fdb721