Update patch set 5
Patch Set 5:

> Patch Set 5: Verified-1
> 
> I am not convinced about 
> 
> .insertshape-grid .row:last-child {
>        min-height: 240px;
> }
> 
> when orientation change to landscape, it does not work well the heuristic value 240px
> if you check the insertshape-grid class style, the position is absolute, so why do not
> look in the code the left, top position are offset a little bit.

I think it is not related to positions. It is related to height of the element. I could adjust its height by using padding or margin, but when i do it, user has to scroll twice to be able to view bottom. Scrolling is not smooth that way. Browser scrolls to start of the margin/padding. User has to scroll one more time to see padding/margin area.

The best solution would be with these properties IMHO:
* Position: relative. => For the row and container.
* Height: auto. => For the row and container.
* Height: 30px => For column.

But for some reason, i could not make it work. Now i will work more on it as you suggested. Thank you for your comment:)

Patch-set: 5