Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I reverted this change because it breaks a CPP unit test. This failure was pointed out by Collabora Jenkins multiple times. TDF Jenkins does not run make check, that's why it does not fail.
> 
> Sorry I can't take Collabora Jenkins seriously as it runs against Collabora Office 6.4 and not against master.

Ahh, I forgot about that we are working on different branches. I did not check whether it fails also with core/master. I'll check that, if not then I'll put this patch back.

> Also, do you know if there is a way to run just a single unit test and not the whole test suite?

I don't know a good way for that. The failure output writes out a very long command, but I did not try it so far.

Patch-set: 4