Use CollaboraOnlineWebViewKeyboardManager if available
If loleaflet is embedded in an iOS app that uses
CollaboraOnlineWebViewKeyboardManager, then we can use that to
reliably display and hide the on-screen keyboard. (Such iOS apps are
Collabora Office or Nextcloud.)

This is optional as we can't be sure whether the version of the iOS
app we are embedded in uses CollaboraOnlineWebViewKeyboardManager or
not. If not, work as before. I.e. hope that calling the focus()/blur()
methods of a textarea object will show/hide the on-screen keyboard.

Change-Id: Idddedcb4a83588c622067cdbeadb02ecdbd4fc72