Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So when you mention 2a16f34812cf69bbe0f21b7e7d048fdb3271fa9d , you mean that if somebody confuses the string "broker" in log output with the class name DccumentBroker, which is the only other thing where the term "Broker" exists in the code, they are just stupid`

Not at all! I felt I didn't explain what was going on with having the same suffix. In fact it was intentional (to have both these threads with similar names), but without explanation it's understandable that it might seem accidental.

But any common name can be confusing and this is why I explain that the threads have nothing to do with classes in code. Indeed, there are dozens of files and classes that log under those two thread names.

I'm more than happy to change "broker" to something else, but the idea is that they are the same so we can filter logs easier (I also made a good attempt at making the log entries align, for easier readability). I hope the advantage of that is clear.

Thanks for reviewing.

Patch-set: 1