Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> As an aside: http://lxr.linux.no/linux+v2.6.37/include/linux/sched.h#L245 leaves us four characters more to play with for the number; perhaps that's enough but its getting right =)

Yes! I was aware of this limitation when adding thread-name registration (some 4-5 odd years ago). So the names are as short as possible, but still meaningful.

Again, we can change them to make them shorter or more meaningful, but I'd like to maintain the symmetry and utility of making them useful as well. Grepping with the common part and the unique document ID seems like a no brainer advantage, IMHO.

Patch-set: 1