Update patch set 2
Patch Set 2:

FWIW - I would expect us to un-conditionally soft-link, and then copy the files as/when we do to this new location the links point to - that is the same as the one used with the new mounting setup (I hope) - to increase commonality and testability.

I'm not at all sure that hard-linking is a good idea; it really depends how the system re-writes those /etc/foo files - quite possibly they are atomically replaced with a 'rename' (that is how I would do it) in which case the hard-linking is not going to do us any real good (then again neither will copying unless we listen / poll for changes to those files ourselves).

Patch-set: 2
CC: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>