Make comments match reality
"Private" and "public" have no meaning in JavaScript as far as I know,
there are only naming conventions. And we keep using properties and
functions intended to be "private" (have names that start with an
underscore) quite freely across source files and classes anyway.

Also, the names in the comments were wrong.

Change-Id: Ib5172c26be8818cb261dbe7d08d1c3245ff64af8