calc canvas: Fix occasional off-by-one error that results in a blurry canvas.
The core of the fix is in _getNewPixelOrigin() where the round() behaves
non-predictably / inconsistently with the rest of the code, causing
random off-by-one error that shows (or not) depending on the window
size.

The biggest problem of this is that this off-by-one is then multiplied
somewhere by the zoom factor, causing the canvas being completely
blurry; but eventually when the user clicked into the sheet, it
'magically' fixed itself.

The rest of the changes (in setZoom()) should actually do the same thing
as the previous code, but using existing methods, instead of computing
the shifts manually.

Change-Id: If0ecb1301b7c1e65cfe8126385ef959c584c5d16