cypress: NC: delete exisitng test file more effectively.
We could not remove the existing test file, if there are
a lots of files in the NC storage, because in this case
not all files were displayed on the screen and so we could
not know whether the file was there or not.
So instead we remove all files, what we can do easily.

Change-Id: Iae7a8ee5c2064ecd94c29d990b7507001827b0dc