Make the 'modifypage' button work on first tap in iOS app
Horrible hack: Could not figure out the root cause, so just pretend
the user tapped twice then the first time.

This is not intended to fix tdf#136457 though. But it might still have
some effect on that whole can of worms. That bug is hard to reproduce
or even descibe exactly, so it is hard to say.

Change-Id: I147a448cceb0adc25fa5d136d59bfebf91839aaf