Update patch set 7
Patch Set 7:

Mert: I've read the code now, looks great in general.  Only thing to improve is that some of the functionality (like onClick in requestForCopy()) is now a copy of createNewFileInputDialog() from LibreOfficeUIActivity.java; can we have just one instance of that in LOActivity, and call that from LibreOfficeUIActivity, to avoid copy/paste of the code please?

Otherwise - let's get it in! :-)

Patch-set: 7
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>