Update patch set 1
Patch Set 1:

The second '*' parameter to that postMessage() call is not needed; this is a different postMessage() method than the one used to post messages to another iframe.

Also, we probably should also remove the re-definition of window.open() from loleaflet.html.m4, otherwise we will open the hyperlink twice in the mobile apps, I think.

Patch-set: 1
CC: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>