Update patch set 1
Patch Set 1:

Hi Szymon; of course - if a document is not actually modified we should work -really- hard to avoid saving / over-writing that in the storage. In theory that can cause data-loss in edge cases. So - I wonder - why we try this especially hard after an hour (?) is that just when we shut the document if it has not been used ? (and perhaps we do that normally when connections are lost or ? ...).

Thanks !

Patch-set: 1
CC: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>