Update patch set 1
Patch Set 1:

>  So - I wonder - why we try this especially hard after an hour (?) is that just when we shut the document if it has not been used ? 

If document was not used for more than an hour (or other defined time) we try to force autosave the document and then close it. This happens without any time delay between trials - what causes CPU usage I think.
In case I analyzed the last document save to storage failed (but internal one not) and every time we try to autosave we fail with "unmodified" state from internal save, but we don't continue with save to storage.

so maybe I will add another condition: force save to storage even on unmodified when the last save to storage failed?

Patch-set: 1