Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think from now on you should use GitHub directly 😊
> > The new location of the repo is: https://github.com/CollaboraOnline/online
> 
> Done https://github.com/CollaboraOnline/online/pull/208 😊

GitHub Collabora repository is not the same as TDF LibreOffice repository, and is only for Collabora Online (so, a product not managed and released by TDF). If you commit there, you are not helping the development of LibreOffice but only of Collabora Online, which is a different product.

Patch-set: 1
CC: Gerrit User 1000009 <1000009@6622c302-437a-49ad-a91d-55ba7dcf1615>